Consistently bind all joypad devices to actions
Previously, some actions had bindings that matched buttons on any joypad (such
as move_up, repel), while others only matched device 0 (such as running,
dialogue_next, ui_accept). This meant you couldn't fully control the game with
any connected joypad. In particular, if device 0 is an Xbox controller while
device 1 is a PlayStation controller, using the dpad on the PlayStation
controller caused the "run" hint in the stealth level to change to the
PlayStation icon, but that button did not actually work on the PlayStation
controller!

This is a single-player game, with no way to choose a controller. So, accept
inputs on all connected controllers for all actions.