ci: Don't fetch all LFS objects for Flatpak build
The Flatpak build needs the icon, which is a PNG, and so is stored in
Git LFS. But it doesn't need the other tens of megabytes of assets,
which take time to fetch and also count against our monthly Git LFS
bandwidth quota.

It is possible to tell Git LFS to only check out certain large files,
rather than all LFS objects referenced in the working copy. But there is
an easier way: we can use git's sparse-checkout feature to only check
out the linux/ directory (plus the files in .) for this job. Then Git
LFS will only fetch the objects used in that directory.

To make this work we have to change linux/icon-256.png from a symbolic
link to a copy of the icon from assets, but I think that is a price
worth paying.