LICENSES: Replace MPL symlink with copy of license
We want to have the project license at the top level of the project
because this is treated specially by GitHub. But we also need to have a
copy of the license in LICENSES to placate `reuse lint`.

Previously the latter was a symbolic link to the former, but I have
remembered that symbolic links are problematic on Windows.

Use a copy of the license instead.