Void grappling: Dress the level
Replace trees by set of bushes. And add bushes all around. Violet ones
at the beginning, and red/violet/blue sets later.

Add rocks in the water too.

Add square patches to the floor.

Also, adjust the position of the last pin on the go. So it is easier
to hook.

https://github.com/endlessm/threadbare/issues/1505