Switch background music when ink combat completed
Previously, the intense combat music would continue to play even after
the final round had been completed and the collectible revealed.

Define an AudioStreamInteractive with 2 clips, the combat music and the
bed used in Fray's End. Set this as the background music stream on all 6
ink combat scenes, rather than setting the combat music directly: this
ensures that the music doesn't restart at the beginning between rounds 1
and 2 of combat.

Then, add a method to background_music.gd to switch to a different clip.
Connect to the goal_reached signal and, in addition to playing the
animation that reveals the collectible, also switch the music to the
the "cozy" clip.

This doesn't work quite right when switching between
quest_003/2_ink_drinker_levels/ink_combat_round_6.tscn and
frays_end.tscn. frays_end.tscn uses the less-dramatic bed music directly;
ink_combat_round_6.tscn uses a AudioStreamInteractive. background_music.gd
treats these as different streams and so restarts the music when switching to
frays_end.

But it sounds OK and I think the trade-off is worth it. I think this could be
improved in future by going further down this path: put *all* the background
music into an AudioStreamInteractive, with appropriate fades; or even have our
background music logic build an AudioStreamInteractive dynamically.