WIP: insect hint for musician puzzle
In my original prototype, the musician demonstrated the first melody,
and then you had to work out the other melodies based on some
crudely-drawn signs. After several iterations we reduced it to its
current form: you interact with an altar and it shows you the melody; or
you repeatedly ask the musician and he tells you what to do.

I would like to restore some of the requirement to think/observe to this
puzzle.

The idea here is that an insect (implemented here as a pumpkin) sits on
a nearby bush. When you disturb it, it flies onto each rock in the
current sequence in turn, triggering a wobble when it touches each rock,
before returning to its bush. I tried to do this with PathWalkBehavior - setting
up the path as a series of arcs, and using the pointy_path_reached signal to be
notified when it touches each rock (and so the wobble should be triggered). I
know I could do this with collisions but I was curious if this would work. It
doesn't: pointy_path_reached seems to be emitted emitted multiple times for a
single point so the wobbles are not in sync with the motion: sometimes when the
insect hits one rock, several rocks wobble.

I didn't spend too much time debugging here - I think what I'd do in practice is
to make the insect/pumpkin collide with the interaction area of the rocks &
pause its motion for the duration of the wobble that way.