Template elevation tileset: Improve for easy terrain
Remove the tiles of walls. Extend the floor tiles that were on top of
the removed tiles to occupy the removed area. Finally, set the rendering
texture origin of those tiles to -32 on the Y axis.

Leave a single "Elevation" terrain. Setup the terrain bitmap to the same
as other floors like grass. So now painting terrains is easy.

Also:
- Adjust the physics layer 0, mapped to "walls".
- Adjust the physics layer 1, mapped to "sight occluders" (for stealth).
- Adjust the occlusion layer for the lights.

And adjust the tileset_test.tscn scene by removing the now inexistent
tiles.

Thanks to @anonomity (Jackie Codes) for her talk in GodotFest 2025,
where she presented about this way of making cliffs in tilesets.

https://github.com/endlessm/threadbare/issues/1613