Set challenge start scene when a thread is collected
The intended behaviour is that when the player runs out of lives, they
return to the start of the current challenge. Collecting a thread marks
the end of one challenge and the beginning of the next. So after the
player collects a thread, set the challenge start scene before changing
to that scene.