interact_input: Remove unused variables
The pressed state was originally handled differently depending on whether the
current device is a keyboard or not, which was tracked by is_keyboard_mode. But
after some other changes to this script, both branches of the if statement are
the same. Remove the variable and all uses of it.

current_device was set to the new device each time it changed, but never read.
Remove it.