Simplify finding textures for directional input
Rather than a global singleton, use a resource and reference it in the
relevant script. This is static data. This also means that we do not
need to use ResourceLoader.load() at runtime: we can lean on Godot's own
resource loading machinery more directly.

Rather than reimplementing device name recognition, use the device
identifiers provided by the InputHelper addon that we rely on as keys
for the dictionary.

Allow the game to crash if the device identifier is not in the map.
We'll find out soon enough.