Rename input hint texture resources to match XML names
The individual assets in the input prompt atlas have canonical names that are
used in the accompanying XML file. Not all of the resource files match the name
of the icon they reference in the atlas. Update the names to match the icon they
reference. This will allow generating these resources automatically, either to
fix coordinates in the existing files or to add new ones.

Remove keyboard_space.tres - it was actually an unused duplicate of
keyboard_space_outline.tres. Rename the others. Update all files that reference them.