Add script to import Kenney asset sheets
As noted in the header, this is derived from
https://bitbucket.org/iseesharp83workspace/kenney-spritesheet-importer-godot-plugin.
I ran into a number of problems with this addon as in that repo today (which is
also in the Godot Asset Library):

1. It assumes that the XML file coordinates use the same coordinate system as
   Godot, but they don't: Godot's Y-axis points downwards while the Kenney
   (Sparrow) XML files' Y-axis points upwards.

2. It embeds the atlas data into each AtlasTexture .tres file, which I later
   found has been fixed by someone else in an unmerged PR:
   https://bitbucket.org/iseesharp83workspace/kenney-spritesheet-importer-godot-plugin/pull-requests/1

I also find it a little weird that it acts as an alternative importer for the
.png file - which means you can't use the .png file directly any more. That's
why in the PR linked above, the author manually creates a Texture2D for the PNG
and saves it as a .tres file.

Ultimately I found it easier to trim the plugin down to an EditorScript that can
be run by hand. In the process I added a tweak to make it less likely to
rewrite identifiers in .tres files for no reason, and added a switch (disabled
by default) that controls whether new textures are created, or only existing
ones updated.