Add aim input hint
Previously, in the first grappling level, we reused movement_input_hints.tscn as
the "aim" hint. For the keyboard this is correct but for gamepads it is not: we
use the right stick to aim. (You could also imagine that we might recommend
using the mouse to aim when moving with the keyboard, some day.)

Refactor directional_input_hint to support being used for a different family of
directional actions. Assign appropriate textures to each joypad/keyboard
resource. Define a new aim_input_hint.tscn scene. Use it in the grappling quest.

Resolves #1373