shjourney: Convert rainforest sound to Ogg Vorbis
I used this command:

    ffmpeg -y \
        -i "scenes/quests/story_quests/shjourney/2_shjourney_intro/template_intro_components/EffectSounds/amazon-rainforest.mp3" \
        -c:a libvorbis \
        -q:a 5 \
        "scenes/quests/story_quests/shjourney/2_shjourney_intro/template_intro_components/EffectSounds/amazon-rainforest.ogg"

The resulting file is about half the size.

I manually edited the .import file to have the same UID as the MP3.

It is not ideal to take a lossily-compressed MP3 file and apply a second round
of (Vorbis) lossy compression to it. The source file cited in the .license is a
lossless WAV file, but it's several hours long; this is just a clip. To my ear
the recompression sounds fine; it's ambient noise anyway.