shjourney: Convert Icy Glare to Ogg Vorbis
At 9.4 MB this is one of the largest files in the game. It is used for a
30-second cutscene. The Ogg Vorbis file is 3.1 MB, converted with:

    ffmpeg -y \
        -i Otros_componentes/musicas/Icy\ Glare.mp3 \
        -c:a libvorbis \
        -q:a 3 \
        Otros_componentes/musicas/Icy\ Glare.ogg

It's not ideal to convert from MP3 to Ogg Vorbis, applying a second
round of compression, but the source file on itch.io is an MP3. To my
ears it sounds OK and I think it's worth it to save 6.3 MB - 6% of the
size of the packed game.

I manually preserved the UID from the MP3 file.