El Ojo Revelador: Add third-party asset licensing information
Several of these assets require attribution; we need to have this
information to hand so that we can generate/write the credits. The
information is in commit acccaf86cbf4d04f72faab32df7eb1aff78d1748 but it
is useful to have it in the repo proper.

(The submitter of lpc-dungeon-elements is named "William Thompson", and
their username is "William.Thompsonj". My name is also William J.
Thompson but I am no relation.)