splash, intro, title, pause: Use sand tile as background texture
Previously we had a 1920×1080 close-up photograph of fabric as the
background image for these scenes. In the splash screen it is tinted
with the Endless Access blurple colour; in the pause screen it is a
semi-transparent overlay.

This image was a large PNG; when exported (with lossless WebP
compression) it was around 3 MB of the 98 MB .pck file!

We have a set of tileable fabric textures already in the game: the tiles
for the world!

Use the sand tile (without borders) as the background for these scenes.
Apply the same modulation to the splash & pause instances as before.

I initially tried using a TextureRect with an AtlasTexture as its
texture and the stretch_mode "tile". However as documented
<https://docs.godotengine.org/en/latest/classes/class_atlastexture.html>
this doesn't work:

> AtlasTexture cannot be used in an AnimatedTexture, and will not tile
> properly in nodes such as TextureRect or Sprite2D.

While looking at bug reports about this limitation, I stumbled upon a
workaround: use a NinePatchRect, but set the patch margins (i.e. the 8
tiles around the edges & corners) to 0px, leaving only the 64px central
patch which tiles correctly.

Helps https://github.com/endlessm/threadbare/issues/1683