Add utility to recolor cel-shading sprites
Add a shader, a material with that shader applied, and a custom node
with a bit of logic to set the shader parameters in a smart way.

The shader replaces 3 key colors with other 3 instance uniforms. Being
instance uniforms is important so that each sprite gets their own
colors.

The behavior node controls a CanvasItem with the shader applied. It can
derive the high and low shades from the medium shade, and knows a list
of skin colors. But is also flexible enough to use any color.

The reference for the skin colors is an image from a blog post,
linked in the documentation comment in code.

Also, add a test scene with NPCs of different skin tones.

https://github.com/endlessm/threadbare/issues/1714