Import sketches with lossy WebP compression
By default (in our project) images are stored with lossless WebP
compression. This does reduce the size of some PNGs, e.g. Void.png is
612 KB, but in the .pck it previously became a 485 KB texture resource.
(But I did see some files which actually get larger; in any case, I am
surprised to learn that any time and effort that we spend crushing PNGs
has no effect on the size of the exported project, only on the size of
the Git LFS object!)

I think lossless compression is appropriate for our pixel-art assets
(which tend to compress very well: they are low-resolution with a
limited colour palette). However we do have some sketches which are
larger.

Adjust their import settings to use lossy compression. In the resulting
.pck, Void.png becomes 231 KB; Musician.png goes from 334 KB to 138 KB,
and the 5 intro sketches go from 1.5MB to 948KB in total.

Resolves https://github.com/endlessm/threadbare/issues/1683