renya: Add license & special permission for music
The standard license for this music does not permit use in an
open-source game. Happily the composer has given us special permission.

Add the standard license for the music, and the special permission, to
the repository. Add a REUSE.toml file to mark all these tracks as
covered by that license.

Resolves https://github.com/endlessm/threadbare/issues/1710