cat: Flick tail about once per second
The cat has a 5-frame animation of it flicking its tail. I think it
looks better if the tail only flicks intermittently.

Increase the duration of the first frame of the animation to 800ms. This
makes the full animation 1200ms, of which only the last 400ms (500ms if
you count the return to the first frame) are animated.