shjourney: Compress tilesetCombate final.png lossily
In the repo this is 1.4 MB; when exported with lossless WebP compression
the .ctex resource is 1.1 MB.

This image has the appearance of pixel art, but zooming in each apparent
pixel is actually an 8×8 block with lots of fine texturing.

Enabling lossy WebP compression results in a .ctex of 14 KB: negligible.

Resolves https://github.com/endlessm/threadbare/issues/1684