Mothsache: Consistently connect signals in scene
Previously, most signals were connected in the scene file; the `_ready`
function also connected them, but only if the callback is not already
connected.

In my experience, not knowing whether a callback is connected or not is
error-prone. The fact that four of them were connected in the .tscn, but
one was not, supports this theory!

Add the one missing connection to the scene file. Remove the
is_connected() checks & connect() calls from the script.