dependabot: Group updates for actions/* namespace
It is common for GitHub to release new versions of all their actions in
near-lockstep, causing a flood of PRs in each of our projects. For example:

https://github.com/endlessm/threadbare/pull/1732
https://github.com/endlessm/threadbare/pull/1733
https://github.com/endlessm/threadbare/pull/1734

were all opened simultaneously, when the weekly dependabot job ran, and they all
perform the same task: update an action to use a newer Node.JS version.

We can't help the fact we have multiple projects, but having read the
documentation it looks like we can at least group these updates together.