RandomFrameSpriteBehavior: Randomize frame progress too
Previously this behaviour would randomize the starting frame of the
parent AnimatedSprite2D. But this can still lead to unwanted
synchronicity between instances of the sprite. For example, in Fray's
End we have a big field of barley. The barley animation has 2 frames
(actually it has 4 but it's 2 unique frames each duplicated once), and
so there are only 4 possible starting positions, leading the field to
have a quite visible lock-step sway.

As well as setting the `frame` property, also set `frame_progress`,
which is the position within the chosen frame (between 0.0 for the start
and 1.0 for the end).

In addition, reset both `frame` and `frame_progress` when saving the
containing scene. (Setting `AnimatedSprite2D.frame` implicitly sets
`AnimatedSprite2D.frame_progress` to `0.0`.) Previously we only cleared
`frame_progress` which means that, for instance, `cat.tscn` has a
particular `frame` saved.