Void runner: Become the forest a collider
By changing the node type to StaticBody2D. Just like the
StartingIslandTrees nodes.

Also adjust the polygon shape by smoothing the corners, so the player
collides nicely with it.

The polygon should potentially continue be used for refilling the area
with trees. But making its shape smoother and preventing hard edges as
it fades into the water was more important here, than keeping a good
shape for refilling. And adding 2 polygons (one for colliding, one for
filling) would be a bit confusing.

https://github.com/endlessm/threadbare/issues/1753