credits: Add community contributors
This includes, to the best of my ability based on the Git commit history, all
contributors to any file outside of scenes/quests/story_quests. The basis for
not including that folder is that StoryQuests have their own credits, shown in
the storybook.

Some people are credited by a screen name/username because I was unable to
determine another name from them based on a quick look at their GitHub profile.

We will of course accept contributions to add, change, or remove names from this
list if people want to be credited differently or if I have missed someone!

Resolves https://github.com/endlessm/threadbare/issues/1036