El Juguete Perdido: Remove Marcos Cabaña from credits
At the time this StoryQuest was submitted, I was proposing to list
mentors in the credits for each StoryQuest, which is why I asked the
creators to credit Marcos here. But no other StoryQuest has a credit for
the mentor.

Now, we have a dedicated section in the credits for mentors.

Remove Marcos' (misspelled) name.