tiles: Add single-tile foam TileSet
The previous foam tileset had five tiles, arranged in a cross shape. In
the diagram below, f is foam ripples, F is the central full-blue square,
and · means no tile. Each character represents 32×32px (i.e. a
quarter-tile) for reasons that will later become clear:

    ··ff··
    ··ff··
    ffFFff
    ffFFff
    ··ff··
    ··ff··

This arrangement is hard to use because when two or more coastlines
border the same tile, you need to have 2 or more foam ripple tiles on
top of each other to get the correct effect. And if you look very
closely at the sprite sheet, there is actually the odd pixel of foam in
the inside corners of the ·.

https://pixelfrog-assets.itch.io/tiny-swords/devlog/1138989/tilemap-guide
explains that this is supposed to be used in a different way. The 64×64
tile should use a 128×128 texture, centred on the solid-blue square. In
this diagram, T is the tile proper, and t is the extra texture region to
draw around it:

    ······
    ·tttt·
    ·tTTt·
    ·tTTt·
    ·tttt·
    ······

With this configuration you place the tile under the coastline, rather
than on top of the water, and the foam animation is shown on the four
neighbouring tiles. The ones that are under the ground are obscured.
Then if you have 2-4 coastlines touching the same tile of water, the
correct effect can be achieved with a single foam layer.

Resolves https://github.com/endlessm/threadbare/issues/1621