Ink combat 3: Improve music transition
Previously the transition between the combat music and the cozy bed when
you complete the challenge was a 2-second crossfade. The problem with
crossfading to the cozy bed is that it has a strong chord on the
downbeat; by crossfading we miss it.

Add BPM and time signature information to the combat music (and to the
other pieces in this suite for good measure). These are by ear, but I
think they're right.

Then configure the combat -> cozy transition to occur at the end of the
current bar, then start the cozy clip at full volume and fade out the
combat clip for 2 beats. (I would prefer it if the FADE_OUT mode meant
"fade out the current clip for 2 beats, *then* start the new one"; it
doesn't, but actually this sounds OK when aligned to the start of a new
bar of the combat music.)

Use a ClipSwitcher rather than the old approach of binding to
BackgroundMusic.switch_to_clip with the extra call argument &"cozy". I
think it's much clearer to have the clip visible in the inspector.