tiles: Add shadows TileSet
This uses the shadow tile from Tiny Swords. Like foam, this is a single
oversize tile that should be placed underneath each cliff-face.

Unlike foam, this doesn't quite work correctly because the shadow is
semi-transparent, and if you have two next to each other horizontally, they
overlap slightly, so there is a darker triangle of shadow where the two cliff
tiles meet at the ground. However it's very subtle and you have to be looking
very closely to notice it; and the net improvement from adding shadows is worth
it.

Resolves https://github.com/endlessm/threadbare/issues/1760
(though we will still need to gradually actually add shadows to each scene).