Add non-CC0 Tiny Swords free pack
To assets/third_party. Add current version as downloaded from itch.io,
which was updated yesterday.

Add a custom license for it, with the explanation that we have been
granted the distribution of this asset pack in Threadbare.

Add a REUSE.toml file to mark all PNG and Aseprite files within the
subdirectory as licensed by the new custom license.

Update the document on licenses to not mention Tiny Swords anymore as an
unsuitable license.

Resolve https://github.com/endlessm/threadbare/issues/1781