Void + grappling: Add second round
Add more content for the mix of Void + Grappling hook mechanics. This
time, adding needles for connections and the longer string powerup. The
player reaches a place with a river that can't be crossed with the
normal hook string length. When the player picks the powerup, the camera
zooms out, and the Void enemy starts chasing. But now the player can
jump to the other side of the river. After which there are islands like
in the first round, but in this case they are further apart, and at
least one needle connection is needed to reach the exit.

Also change the teleport of the original void_grappling to switch to
this new scene, and add a teleport to go to the outro from this one.

https://github.com/endlessm/threadbare/issues/1349