Add character randomizer script
A single button to randomize various aspects of a character. For now,
the skin tone and the textures of a stack of AnimatedSprite2D that
compose the character with parts like: legs, body, head, hair, head
prop.

Also provides a way to flip the character horizontally. And a way to
start the current animation of each part at a random frame, but all
using the same random frame. The caveat of having multiple parts is that
existing logic like the RandomFrameSpriteBehavior don't work for an
array of sprites, so some things have to be reimplemented.